Start Seed and example index product
