+++ lo-mercat-app/docker-compose-next.yml
@@ -8,3 +8,5 @@ services:
       - adminjs
     ports:
       - "3000:3000"
+    environment:
+      DATABASE_URL: postgresql://postgres:${POSTGRES_PASSWORD}@db:5432/initdb
