+++ lo-mercat-app/docker-compose.yml
@@ -8,6 +8,8 @@ services:
       - db
     ports:
       - "3001:3001"
+    environment:
+      DATABASE_URL: postgresql://postgres:${POSTGRES_PASSWORD}@db:5432/initdb
   db:
     image: "postgres:latest"
     environment:
