+++ lo-mercat-app/package.json
@@ -6,7 +6,7 @@
   "version": "0.1.0",
   "private": true,
   "prisma": {
-    "seed": "node prisma/seed.ts"
+    "seed": "node prisma/seed.js"
   },
   "scripts": {
     "dev": "next dev",
